+++ tests/Set/PartialIntersectionTest.php
@@ -1209,4 +1209,30 @@ class PartialIntersectionTest extends \PHPUnit\Framework\TestCase
         // Then
         $this->assertEqualsCanonicalizing($expected, $result);
     }
+    /**
+     * @test partial intersection of disjoint fresh objects is empty
+     */
+    public function testIsEmptyForDisjointFreshObjects(): void
+    {
+        // Given
+        $left = (static function (): \Generator {
+            for ($i = 1; $i <= 6; ++$i) {
+                yield (object) ['value' => $i];
+            }
+        })();
+        $right = (static function (): \Generator {
+            for ($i = 1; $i <= 6; ++$i) {
+                yield (object) ['value' => $i];
+            }
+        })();
+
+        // When
+        $ids = [];
+        foreach (Set::partialIntersection(2, $left, $right) as $value) {
+            $ids[] = \spl_object_id($value);
+        }
+
+        // Then
+        $this->assertSame([], $ids);
+    }
 }